requirements.txt
gradio
google-generativeai
gtts
python-dotenv
